Added lua scripting
Added a big update to use lua scripts to define parts of the renderer.
Only lights and a single model is supported so far.

Need to support multiple models and have callbacks in the lua file to
update positions and all

NOTE: this has not been compiled for the Dingoo yet, it is very likely
not to compile properly